fix(remote-root): refresh planet status against the live local planet
The remote-root planet badge classifies against the controller's current
custom planet hash, but planetStatus was cached in the DB and only
recomputed on manual health checks / planet mutations. The 5-minute
CheckRemoteRoots cron refreshed remotePlanetHash yet never re-ran the
classifier, so after regenerating the local root list the badge stayed
stuck on a stale CUSTOM_MATCH ("planet 最新") even though the local hash
had changed.

Extract getLocalCustomPlanetHash + the live classifier into a dedicated
remoteRootLocalPlanetService (keeps the pure classifier module free of
ZT_FOLDER/prisma side effects so its unit test stays isolated) and have
the cron recompute planetStatus on every read, matching the router path.
